requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
playwright>=1.40.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
selectolax>=0.3.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0

